backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.50.0
sse-starlette>=2.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
# Authentication
bcrypt>=4.1.0
PyJWT>=2.8.0
# Validation
pydantic>=2.5.0
email-validator>=2.1.0
jsonschema>=4.20.0
# Caching
redis>=5.0.0
# Security
cryptography>=42.0.0
